requirements.txt
requests
python-dotenv
langchain
langchain-core  # Add this
langchain-community  # Add this
langchain-openai
langchain-community

openai
chromadb
sentence-transformers
pandas